Allow user to control how to interpret voltage data captured in BUF
This is done using --dtype and inturn "b" or "B" to identify
signed or unsigned 8 bit data.

Dont use h or H or ... ie any types which will be interpreted as
larger than 8 bit data, bcas it will be wrong wrt the actual
data stored and thus will exhaust the data before the logic is
done with extracting the data.

NOTE: The default type has change from unsigned to signed here