Plot the Channel's Base line
NOTE: The YPos wrt the Channel seems to not align with the sample
data of the channel AND NOR with the position it should be in
a virtual screen space which duplicates the device.

Based on how the yposition was setup on the device screen, and
what I am seeing corresponding to it in the dumped buf file,
it seems to require a 1.28 multiplying factor to align with the
virtual screen space.

* screen split into 8 grid blocks vertically. And with ypos
  set at the boundry between 1st-2nd blocks, assuming a 256 pixel
  vertical screen space, the value needed is 32, but what is found
  is 25. So x1.28 correction factor. This correction factor also
  seems to inturn apply equally well when ypos set to boundry
  between 6th and 7th blocks.

Parallely the data samples seem to require some other multiplying
factor to align in the virtual screen space with how it should be.