Plot Voltages wrt Y Axis - Single Channel plots only
NOTE: A given plot can have only channel being shown. Else if more
than 1 channel is shown, the voltage corresponding to last channel
in the list will be used wrt voltage plot. Rather need to change
logic so that last analog channel in the list should determine
the voltage mapping wrt y axis ideally. ie 0 or 1 and thus if
both 0 and 1 is being shown, then use 1's voltage mapping or may
be give priority to channel 0 ie channel a.

Multiple channels and Voltage handling needs to be handled properly
in the flow later.

MAYBE we need to provide either

* have each channel be in its own subplot

* have all channels on the same plot, but then dont show voltage
  wrt y axis, instead just mention voltage range wrt each channel
  as a annotation.