Count number of ups and downs on the way to calcing freq
Count the number of ups and downs in the waveform between the
two clicked x positions on the waveform.

The logic tries to find a upward moving path of the signal first,
so click towards the bottom of the waveform, so that a downward
moving part of the waveform that will be ignored, if found 1st,
is handled / worked around / avoided.