A variation of midi test script
Avoid the sleep btw NoteOn and NoteOff messages, to see if both
messages are captured. Still only NoteOff seems to get captured
(or is it sent).

Changed from 0x00 to 0x55 wrt velocity byte in NoteOff message,
and the difference is seen in the captured data.

TOCHECK: Check if the change from 0x55 to 0xaa for NoteOff is
captured by DSOQuad Osc or not. It should.

TODO: have only NoteOn messages sent and see if the same is
getting captured. Need to check.