OTDivs: Fix oversight wrt timeAdjust, Default MarkersString with P
Had forgotten to set/reset timeAdjust to 1.0 irrespective of
whether marker strings are still there or not. Without this the
logic will get stuck in a infinite loop, if P which sets timeAdjust
to 0.0 is the last marker in the marker string (either user supplied
or the new default marker string).

Also switched the program's default markers string to use P instead
of previous #print# which has been removed now.