Very skeletal fft plot
NOTE: The sampling rate is hardcoded and not picked from buf file
Need to check if buf file, has the sampling rate stored somewhere
or not.

NOTE: Also the sampling rate is blindly guessed from captured sig
generator data, and not picked from DSO firmware or ideally buf
file.

NOTE: The hardcoded sampling rate is there only for 2 time bases
as of now.