Fft: Oops Me, Calculate SR from timebase and samples/timediv
bcas what we have is not the raw adc sampels, but what is finaly
plotted on the screen, so the number of screen pixels|samples
per time div, will give the sampling rate wrt the data which we
have with us to work on, in the buf file.

Inturn also clean up the show fft logic a bit.

What I had previously infered was same but coming from a self
blinded alley, so some implications of the same have been cleaned
up now.